pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "git-chronicle"
version = "0.1.0"
description = "AI-powered git history narrator — turns your git log into readable stories"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Aaron Wu", email = "faw21@github.com"},
]
keywords = ["git", "ai", "llm", "history", "changelog", "story", "developer-tools"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Version Control :: Git",
]
dependencies = [
    "gitpython>=3.1.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "httpx>=0.24.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.18.0"]
openai = ["openai>=1.0.0"]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "anthropic>=0.18.0",
    "openai>=1.0.0",
]

[project.scripts]
chronicle = "chronicle.cli:main"

[project.urls]
Homepage = "https://github.com/faw21/chronicle"
Repository = "https://github.com/faw21/chronicle"
Issues = "https://github.com/faw21/chronicle/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/chronicle"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=chronicle --cov-report=term-missing --cov-report=html"

[tool.coverage.run]
source = ["src/chronicle"]
